requirements.txt
pytest == 7.4.4
allure-pytest == 2.13.2
selene == 2.0.0rc4
selenium == 4.16.0
webdriver-manager
pytest-playwright == 0.5.2